Check the message detail view with an instance and message with different ids.
I think there's a problem with the message detail view
looking up the message using the writeitinstance_id where
it should be using the message_id.

It also looks as thought a lookup for a message with the right
writeitinstance might fail with more than one message for an
instance, so this test has two messages for the writeitinstance
to spot that if it happens again.