Create management section of urls.
Create a set of management urls for the instance owner.

Move the following to this section:

1) Instance advanced edit
  - This actually manages the WriteItInstanceConfig model, but we get to it
  via the WriteItInstance slug.

2) The API settings configuration
3) Add popit sources page
4) The urls for managing email templates
5) recipients url
6) Add messages section
7) Message moderation

Remove question mark so slash is added to urls - instead of allowing either
with or without a trailing slash, redirect from urls without the trailing
slash to with so that there is a canonical url.