Fix user checking for template editing views.
The urls for the template editing views don't include the primary
key of the appropriate template object, just the slug of the
writeitinstance, which is fine, because there is a one-to-one
mapping between these objects. But it does mean that me can't inherit
from WriteItInstanceOwnerMixin, which expects to see the object's
primary key.

To solve this, don't inherit from WriteItInstanceOwnerMixin in
UpdateTemplateWithWriteitMixin, and implement a custom get_object
(which still checks the owner is correct).