WIP #610. This adds the backend for manually resync with popit
Resyncing a preexisting writeit instance with a popit instance.
This only provides the backend for resyncing. But the frontend would probably look something like this.

In somewhere like ```nuntium/templates/nuntium/profiles/writeitinstance_and_popit_relations.html``` in the
```{% for record in relations %}``` you would have to add something like following:
```
<a class="resync" href="{% url 'resync-from-popit' slug=record.writeitinstance.slug popit_api_pk=record.popitapiinstance.pk %}">resync</a>
```

and at the bottom of the file:

```
<script type="text/javascript">
$(function(){
  $('.resync').click(function(event){
    event.preventDefault();
    var url = $(event.target).attr('href');
    $.post(url, function(data){
      location.reload();
    })

  })
})
</script>
```