WIP #610. This adds periocity to the execution of the task 'update_all_popits'.
The default value is currently weekly.

The possible periodicity values are:
* Never
* Twice every Day
* Daily
* Weekly

I'm thinking that probably we don't need the attribute ```autosync``` and that we can just remove it.

This commit does not add a way for a user to change the values of how often they want their popits to be resynced.