Make the testing output less verbose
This disabled all logging during the tests, which makes it much easier
to see which tests failed. In the future we might want to re-introduce
some logging so that we can see warning and the like. But at this stage
of development it seems more valuable to be able to quickly see which
test needs fixing.