Don't use javascript to submit POST requests
Instead of using javascript to submit POST requests to the server we
use a form. This means that the page works without javascript and we
don't have to worry about the user clicking the wrong target etc.

I've also taken the opportunity to add some flash messages informing the
user what's happening when the click moderate of toggle the visibility
of the message.