Modifiying previous passing test and adding a new failing one
This modifies the test
MessageCreationTestCase.test_go_through_the_whole_message_creation_process
to modify some of the name of the attributes in the case that a form
has prefix (https://docs.djangoproject.com/en/1.7/ref/forms/api/#prefixes-for-forms).

And adds a new failing test to catch the issue in case of changing
instances and not sharing the data across.