Changing wording for re-adding a PopIt.
Was: `You already have this popit instance relate. But we are updating the database anyway, to pull in any new contacts`.
Now: `You have already added this PopIt. But we will fetch the data from it again now.`

(I still don't think it's great, but there's an entire textual overhaul scheduled from Myf on Tuesday...)