Make `name` required in the Widget
Add some client side validation for the Name field.

I can’t work out how to add this for the URL field too. Django seems to
be doing some deep magic on it to make it so that it must look like a
URL, but I don’t know how to add a ‘required’ onto that as well.