Split live sites and test sites
Show the live sites and test sites separately in the Site Manager.

Only test sites should have a Delete button.

It’s important for the Test Sites to get some better explanatory text,
as the New User journey will involve them only have a single site, in
test mode. The current text is just a placeholder that will remind us
to create something better.

I’m not entirely convinced that the views.py implementation of this is
the right way to go about this at all…