Remove double flash after site creation
On the Site Manager we report if we’re actively fetching data for that
instance from PopIt.

However, all admin pages already report on this in the header bar.

So replace this message with the inprogress spinner.

(I suspect we should just be including that line of JS on all admin
pages, rather than pulling it individually like this, but’s one for
another change.)