Disallow going directly to the message page
Even if we don’t provide links to the message composer, someone might
know the URL from other sites, and try going directly to it. So
disallow it there too. (AIUI, trying to go directly to any page further
into the process will redirect back to ‘who’ anyway, so it’s enough to
just check here.)