Refactor Dave’s javascript, and use Bootstrap popover
A pretty major refactoring of Dave’s initial prototype. The Boostrap
.popover() module is used for quick styling and interaction handling,
but it could easily be swapped out for another javascript modal module
in future.

Notably, during the transition to Bootstrap, I removed the ability to
specify a target element into which the help content should be injected.
This simplifies the code, and was only included in the first place
because Dave thought it might make working with Bootstrap popovers
simpler. As it happens, recreating a new popover element on each
click is actually simpler in Bootstrap, so I removed the feature.