Cache jQuery ajax requests to speed up consecutive help-loader hits
We make use of the fact that a jQuery .ajax() call can return a promise,
and that calling .done() on a promise that fired ages ago will still
work over and over, making promises ideal for caching web responses.

This commit also clarifies the two parts of the urlParts array, to
hopefully be a bit more obvious throughout the new bit of code.