Hide admin links to Moderation and Rate Limiting
We don’t want to link to Moderation or Rate Limiting for now as they’re
not ready for public consumption.

The pages will still exist, so that we can tell people individually
(with suitable warnings) that they can use them, but we don’t want to
make them obvious to normal admins.