Add "Disable Site" link to the instance manager
We can use the “Allow Web Form?” option as a way of disabling an
instance — as all we want to do in such a case is stop people writing
new messages. Replies etc should still be accepted, and messages should
still be displayed that already exist. If someone has a strong reason
why a site should completely vanish after it goes live, they’re going
to need to ask a super-admin to do that.