Move moderation to the message admin
Currently moderation happens from the Messages page, rather than the
individual Message page. This is bad, as you want to be able to read
the message before moderating it. So move the button to there.

NB: This has all the same problems as the previous version, but but
it’s a slightly better version of broken here, and helps clean the
Messages Manager up a little bit more for other things we want to do
there.