Only specify one way of running tests
When I first created a new Vagrant box I read the readme and it said to run
`./test.py` but that didn't work. Then I noticed in the motd when you login to
the Vagrant box it says to us a different command. This is confusing so this
commit gets rid of that extra command and just documents the one way of running
tests.

Fixes #1117