Look up instance default language in middleware
If there is a default language for an instance we set that in the
session so that the instance appears in the correct language.

We've changed the order of the middleware here so that
'request.subdomain' is available before we set the locale. This means
that we can pull out the current instance and check it's default locale
using 'request.subdomain'.