Prevent the language from getting stuck in the session
We're using a slight hack to work around the fact that Django doesn't
allow you to hook into the language selection process in middleware. We
check if the user has specified a default language in their session, if
they haven't then we temporarily use that as a place to store the
default lanuage for a session. We then delete our override from the
session so it doesn't persist when the user is viewing other instances.