Fix for tastypie no longer creating API keys for users in fixtures
Since django-tastypie/django-tastypie@7e02e5ffbe226ee756b0666356
(released in 0.13.2) the create_api_key function no longer creates
an API key if 'raw' is truthy in the kwargs. This is correct behaviour
according to the documentation:

  https://docs.djangoproject.com/en/1.9/ref/signals/#post-save

However, it means that there is no ApiKey object created for the admin
user created from the example_data.yaml fixture, since when loading
fixtures raw is set to True.  To get around this problem, this commit
adds an API key for this user in the fixture.