Fix for multiple WriteItInstance errors on API requests
In the previous version of django-tastypie, there was much less checking
done of the kwargs used to get a WriteItInstance from the ID in the URL.
However, in the current version, check_filtering is used to check that
the fields being filtered on are allowed according to the resource's
self._meta.filtering, and since there was none on WriteItInstanceResource
the 'id' filter was ignored and all WriteItInstance objects were
returned.

In order to find the WriteItInstance from the URL, we only need to allow
an 'exact' match on 'id', so that's what this commit adds.