Move some models to a new instance app to avoid circular dependencies
There was an unnecessary circular dependency between the 'nuntium' and
'contactos' Django apps:

   contactos.Contact needs nuntium (for nuntium.WriteItInstance)
   nuntium.OutboundMessage needs contactos (for contactos.Contact)

This wouldn't be a big deal, except that when we're migrating to Django
1.7 it means that makemigrations generates multiple initial migrations
(0001_initial and 0002_auto..., the latter of which adds the problematic
foreign keys). The second of these migrations didn't work properly on
Django 1.7 and SQLite, so the easiest way to deal with that upgrade
seems to be to:

 * Remove the circular dependencies with South migrations and
   Django 1.6
 * Then do the upgrade to Django 1.7, when we should now be able
   to have single initial migrations

This commit largely deals with the first of those by creating a new
'instance' app, which contains:

  WriteItInstance
  Membership
  WriteItInstanceConfig
  WriteitInstancePopitInstanceRecord

... which were previously in 'nuntium'. This commit includes the
migrations to rename the tables, update the content types and the
foreign key references to those models, as well as the updates to
the models files.

The next commit will make the code changes (mostly import fixes)
to go along with these model changes.