Redirects from / to a language-specific variant shouldn't be permanent
It's annoying for local development that the redirect that's returned
from '/' is a permanent redirect (301) since then localhost:8000
redirects to, say, http://127.0.0.1.xip.io:8000/en/ even when working
on other sites, meaning you need to clear your browser's redirect cache
(not simple on all browsers).

It seems wrong that this is a permanent redirect anyway, since the
language prefix on the URL depends on Accept-Language, and you might
change languages in your browser.

This commit changes the redirect to 302 Found; widely interpreted as a
temporary redirect.

Fixes #1170