Freeze versions in requirements.txt
Previously we were just specifying the packages that we wanted, rather
than the specific versions of those packages. This change pins the
packages to a specific version so that when new versions of packages are
released they don't end up breaking things unexpectedly.

I generated this version of requirements.txt by creating a fresh
virtualenv, updating to the latest pip with 'pip install -U pip' and
then installing the requirements.txt and running 'pip freeze' to pin
down the versions.

We should probably look into doing this in a slightly more modern way,
e.g by using pip-tools 1.0 [1]

1: http://nvie.com/posts/pip-tools-10-released/