Fix the exception on moderating a message created with moderation off
This test was introduced in 35f54775889; when that commit was written,
confirming a message created with moderation off didn't cause it to be
marked as moderated.  It thus expected an error about the Moderation
object not existing for that message, and for the message to stay with
moderated == False.

However, cbba85a1dd7b695f7 changed the behaviour of the
'recently_confirmated' method (the one that's called to indicate that a
message has been confirmed) to mark a message as moderated on
confirmation if moderation was off.  In addition, 88e3c97cb4fda685
introduced the change that a ValidationError would be thrown from
moderate() if the message was already moderated. (I don't think this is
the right exception to be throwing for that case, but it follows the
convention in the existing code, so I'm not changing that.)

This test seems to be of limited usefulness since cbba85a, since it's
only in odd situations that moderate() will be called on a message
created with moderated == True, and that's now the case for messages
created when moderation was off. However, I've updated the test for the
current behaviour, by:

  (a) Checking for ValidationError instead of Moderation.DoesNotExist

  (b) Removing the assertion that moderated == False