Fix alignment of <small> parts of the label
The <small> part of the label hint was always appearing to the right,
which was wrong in RTL languages.  Since the display appears to be OK
without the 'float: right', I've removed it.  (The intent seems to be to
have the note right-aligned and the main label left-aligned, but that
wasn't working anyway.)

A better idea in general, which @dracos pointed out, is that we could
have a Sass variable for $left, which might be 'left' or 'right'
depending on the direction, and then a flip function to swap the left
and right parameters to margin, say - see:

  https://github.com/mysociety/fixmystreet/blob/master/web/cobrands/sass/_mixins.scss#L46-L63

In order for the Farsi label to appear in the right order, I've also
had to reverse the two parts of the label in the translations.