Do not display manager interface as RTL
There is a problem with chrome where if the manager is set to RTL
then the page does not display until you resize the window. As the
manager interface currently needs work in RTL layout (see #1184) turning
of RTL display altogether is the best solution.

This is a temporary workaround for #1185.