Replace the broken references to chosen.css
Two template files referenced {% static 'css/chosen.css' %}, which was
removed in 6009225cbb23f46f.  That causes errors in production when
loading pages base on those templates.

That earlier commit switched to using bootstrap's styling for
the chosen.js-enhanced UI elements; this commit does the same to replace
the chosen.css references that are left.