Allow the language code to be an argument to the join_with_commas template tag
join_with_commas inserts an "and" (literally, i.e. not a translation of
"and") between the last two elements of a comma-separated list. This is
only valid for English, so this commit changes the template tag so that
it only uses that rule if the language code is en* or unspecified.