Add the 'uncontactable' class appropriately & fix a performance issue
As an example, it was the people whose names ended in a lower-case vowel
that were displayed as uncontactable.  This commit changes that so that
the uncontactable class is added for the people who don't have any
Contact objects with is_bounced=False.

It also seemed that the widget rendering was causing an extra query for
each option in this list; this commit also fixes that and adds a
regression test for the number of queries used.