Add an instance config option for stricter anonymity
There is a view ('all-messages-from-the-same-author-as') which, given a
message slug, allows you to list any other messages with the same author
email address as the message with that slug. In situations where anonymity
is important, this is undesirable - someone might send a sensitive message
with a false name and an uncontroversial one with their real name, in
which case this view would reveal that they're sent by the same person.

This commit introduces a new instance config option
('email_and_name_must_match') which means that the
'all-messages-from-the-same-author-as' view will only show messages that
have the same author_email *and* author_name, to avoid this information
leak.

Fixes mysociety/alpaca#35
Fixes ciudadanointeligente/write-it#1194