Fix the Vagrant provisioning script's shebang line
The scripts/provision.sh script, which is used for provisioning Vagrant
boxes, used a shebang line of #!/bin/sh, but the script itself uses
bashisms like '[[' for conditionals. This means that the script didn't
work where /bin/sh is a stricter Bourne shell like dash.

So, this commit changes the shebang line to #!/bin/bash so that it works
on more systems.