If we can tell which people are current, only include them
Since the migration to syncing people using django-popolo and
multiple-django-popolo-sources, we're now syncing all people (even
historic representatives) from EveryPolitician, not just those in the
current term(s).  However, we should only be allowing people to write
to current representatives.

The problem here is that the old behaviour was just to include everyone
from the data source who had a contact and hadn't been deliberately
disabled - some (or perhaps most) of these non-EveryPolitician data
sources won't have any way to tell who is current or not; the assumption
should continue that they should all be made available to write to.

We could divide these cases into sources that look like they're from
EveryPolitician or not from the URL, but that wouldn't help with people
serving EP data from elsewhere, or after modification.  Probably the
best approach would be to only show current representatives for sources
that have Event objects with classification "legislative period", and
then decide who is current by looking at those who have memberships with
a legislative_period_id corresponding to a current legislative period.

Unfortunately, at the moment django-popolo does not include the Event
class, and we need a workaround quickly for the Alpaca project, so this
commit uses a different heuristic for when to attempt to filter people
for being current memberships of a legislature: if the people from that
source have any memberships at all, it only includes those people who
have a current membership (of any organisation - of course this might
be completely spurious). This is good enough for the Alpaca case, which
will give us enough time to update django-popolo and
multiple-django-popolo-sources so that we can do this properly for the
main production writeinpublic site.

This commit also doesn't address whether historic representatives are
available via the API.