Preserve legacy identifiers when syncing from a data source
The default behaviour is to update identifiers to remove any not in the
upstream data source when syncing, but we've deliberately created some
legacy IDs in migration that we want to keep.

(n.b. the popolo:person scheme is used is implicitly preserved because
we specified an id_prefix of 'popolo:')