Django 1.8: fix two tests of exception handling when sending email
It looks as if django.core.mail.mail_admins now uses
django.core.mail.EmailMultiAlternatives.send whereas it didn't in
earlier versions of Django. mail_admins is used in the handling of an
exception, so we need to patch that to stop our test exception being
raised in the exception handling as well.