mySociety deployment: changes for timestamped deploy
Our deployment system can now support lower downtime deploys where
everything except migrating the database happens while the site is still
up, being prepared in a new deployment directory whose name is based on
the current timestamp. To take advantage of this, we need to switch the
virtualenv and STATIC_ROOT to be a directories within the repo, and
split the deployment script into two.